requirements.txt
openai
duckduckgo-search
httpx
beautifulsoup4
python-dotenv
rich
